Cast pointer to long instead of int
Matches the cast passed in ((pointer)(long) server->index on line 1276)
Fixes gcc error:
di/wire.c:1119:27: error: cast from pointer to integer of different size

+++ di/wire.c
@@ -1116,7 +1116,7 @@ ProxyWorkProc(ClientPtr unused_client, pointer index)
     XServerPtr  server;
     xLbxAllowMotionReq req;
 
-    if ((server = servers[(int)index]) == 0)
+    if ((server = servers[(long)index]) == 0)
 	return TRUE;
     if (!server->initialized)
 	return TRUE;
